added support for .forced-* classes
In the responsive grid layout, when a component has any parent component
with class forced-xs / forced-sm / forced-md or forced-lg, then the
component behaves up to this forced-class instead of actual
(full-screen) viewport.
2 levels of components nesting supported (a component with its own
responsive grid positioning inside a part of another component with own
grid.
Suported "forced" features are:
- col-*-width
- col-*-offset
- col-*-push & pull
- visible-* & hidden-*

All code is added to *.less files, so for use in app a bootstrap build
of css is needed.
There are html files in the TestForced directory for testing.